add Professor and Students view:

See merge request foxstudent104853/universitycms!8

+++ src/main/java/ua/foxminded/pskn/universitycms/dto/StudentGroupDTO.java
@@ -13,7 +13,7 @@ public class StudentGroupDTO {
 
     private Long groupId;
 
-    private int facultyId;
+    private Long facultyId;
 
     private String groupName;
